geoserver/geoserver/settings/s3utils.py
'''
Created on Jul 22, 2014

@author: minjoon
'''

from storages.backends.s3boto import S3BotoStorage

StaticS3BotoStorage = lambda: S3BotoStorage(location='static')
MediaS3BotoStorage = lambda: S3BotoStorage(location='media')