requirements.txt
MySQL-python=1.2.5
unipath=1.0
django-storages
boto